Move tailor interfaces to package
