+++ apps/frontend/components/editor/Sidebar/ElementSidebar/ElementMeta/Relationships/index.vue
@@ -14,7 +14,8 @@ import type {
   ContentElement,
   Relationship,
   RelationshipType,
-} from '@/api/interfaces/content-element';
+} from 'tailor-interfaces/content-element';
+
 import RelationshipItem from './RelationshipType.vue';
 import { useContentElementStore } from '@/stores/content-elements';
 import { useEditorStore } from '@/stores/editor';
