+++ apps/frontend/composables/useContentElementBus.ts
@@ -1,6 +1,5 @@
+import type { ContentElement } from 'tailor-interfaces/content-element';
 import { getElementId } from '@tailor-cms/utils';
-
-import type { ContentElement } from '@/api/interfaces/content-element';
 
 export const useContentElementBus = (element: ContentElement) => {
   const { $eventBus } = useNuxtApp() as any;
