+++ apps/frontend/lib/revision.ts
@@ -1,12 +1,12 @@
 import { lower, title as toTitleCase } from 'to-case';
+import type { Activity } from 'tailor-interfaces/activity';
 import { assessment } from '@tailor-cms/utils';
 import get from 'lodash/get';
 import isEmpty from 'lodash/isEmpty';
 import reduce from 'lodash/reduce';
+import type { Revision } from 'tailor-interfaces/revision';
 import { schema } from 'tailor-config-shared';
 
-import type { Activity } from '@/api/interfaces/activity';
-import type { Revision } from '@/api/interfaces/revision';
 
 const describe = {
   REPOSITORY: describeRepositoryRevision,
