+++ apps/frontend/package.json
@@ -22,6 +22,7 @@
     "eslint-config-prettier": ">=9.1.0",
     "nuxt": "^3.12.2",
     "sass": "^1.77.6",
+    "tailor-interfaces": "workspace:*",
     "vue": "^3.4.29",
     "vue-router": "^4.3.3"
   },
