+++ apps/frontend/pages/admin/user-management.vue
@@ -105,10 +105,10 @@
 <script lang="ts" setup>
 import { formatDate } from 'date-fns/format';
 import humanize from 'humanize-string';
+import type { User } from 'tailor-interfaces/user';
 
 import { user as api } from '@/api';
 import { useAuthStore } from '@/stores/auth';
-import type { User } from '@/api/interfaces/user';
 import UserDialog from '@/components/admin/UserDialog.vue';
 
 definePageMeta({
