+++ apps/frontend/stores/activity.ts
@@ -3,12 +3,12 @@ import {
   calculatePosition,
   InsertLocation,
 } from '@tailor-cms/utils';
+import type { Activity } from 'tailor-interfaces/activity';
 import { Activity as Events } from 'sse-event-types';
 import findIndex from 'lodash/findIndex';
 import Hashids from 'hashids';
 import { schema } from 'tailor-config-shared';
 
-import type { Activity } from '@/api/interfaces/activity';
 import { activity as api } from '@/api';
 import sseRepositoryFeed from '@/lib/RepositoryFeed';
