+++ apps/frontend/stores/content-elements.ts
@@ -1,8 +1,8 @@
 import { calculatePosition } from '@tailor-cms/utils';
+import type { ContentElement } from 'tailor-interfaces/content-element';
 import { ContentElement as Events } from 'sse-event-types';
 
 import { contentElement as api } from '@/api';
-import type { ContentElement } from '@/api/interfaces/content-element';
 import sseRepositoryFeed from '@/lib/RepositoryFeed';
 
 type Id = number | string;
