+++ apps/frontend/stores/repository.ts
@@ -1,9 +1,10 @@
-import { repository as api, tag as tagApi } from '@/api';
 import type {
   Repository,
   RepositoryUser,
   Tag,
-} from '@/api/interfaces/repository';
+} from 'tailor-interfaces/repository';
+
+import { repository as api, tag as tagApi } from '@/api';
 import repositoryFilterConfigs from '~/components/catalog/Filter/repositoryFilterConfigs';
 import { useAuthStore } from './auth';
