+++ packages/tailor-interfaces/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "tailor-interfaces",
+  "type": "module",
+  "version": "0.0.1",
+  "description": "Tailor data interfaces"
+}
